Version 0.3.0 The ANIMATION PLAYER is now in Alpha! (documentation may take a while to catch up)
-- Animation Player:
The animation player is out in Alpha!
Documentation is minimal so you'll have to make do with the comments in the squapi file and the little bit of documentation I got on notion.
Though don't be afraid to ask me how something works if you don't get it since I like helping out with that
It is prone to errors and mistakes(alpha duh), so if you get something like that make an issue on the github so I can keep track or else I'll forgor
It is also prone to changes so be aware of that when updating

-- Additions:
- Better Error Mentions courtesy of @jimmyhelp!
- squapi.randimation() object added! A simple class that allows you to create an animation to randomly play(like squapi.blink()) as well as modify

-- Changes:
- float Point overhaul! It is now object-oriented so it'll work slightly differently from the other functions, but it will allow for more manipulation :)
- squapi.floatPointOld() is a temporary legacy version if you need, though it will be removed in the future.
- squapi.blink() updated so that it doesn't run when you sleep!
-

-- Fixes:
- fixed issue with eyes when looking down
- fixed issue with tail causing it to tweak out when moving